+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- (REVIEWS-107) Translate 'Migrate Data' button to all languages.
+
 ## [3.10.0] - 2022-07-14
 
 ### Added